pyproject.toml
[tool.poetry]
name = "marlin-cli"
version = "0.0.13"
description = ""
authors = ["George Cooper <george@marlincode.com>"]
readme = "README.md"
packages = [
{include = "marlin_cli", from = "./"},
]

[tool.poetry.dependencies]
python = "3.11.1"
click = "^8.1.3"
gitpython = "^3.1.30"
requests = "^2.28.2"
python-dotenv = "^0.21.1"
pre-commit = "^3.0.4"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.7.0"
black = "^23.1.0"
pre-commit = "^3.0.4"

[tool.poetry.scripts]
marlin = "marlin_cli.marlin:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
